backend/requirements.txt
# VBS Production Management Platform - Backend Requirements
# Python 3.10+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS & Security
python-dotenv==1.0.0

# Data Handling
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2

# Development
pytest==7.4.4
httpx==0.26.0

